refactor make_trace_groups to more general lib function instead of cartesian-specific
