+++ src/lib/index.js
@@ -118,6 +118,8 @@ lib.clearThrottle = throttleModule.clear;
 
 lib.getGraphDiv = require('./get_graph_div');
 
+lib.makeTraceGroups = require('./make_trace_groups');
+
 lib._ = require('./localize');
 
 lib.notifier = require('./notifier');
